significant simplification to zero-values
