+++ README.md
@@ -1,4 +1,16 @@
-# Simplified pure Go secure shell (SSH) client
+# Simplified pure Go secure shell (SSH) multi-host client
 
 [![GoDoc](https://godoc.org/github.com/rwxrob/ssh?status.svg)](https://godoc.org/github.com/rwxrob/ssh)
 [![License](https://img.shields.io/badge/license-Apache2-brightgreen.svg)](LICENSE)
+
+## Test cases
+
+### RunOnAny: all up
+
+### RunOnAny: one down temporarily
+
+### RunOnAny: one down permanently
+
+### RunOnAny: two down
+
+### RunOnAny: all down
