--- doc.go
@@ -1,4 +1,0 @@
-/* Package ssh is simplified encapsulation of the standard x/crypto/ssh
-* package with reasonable defaults
- */
-package ssh
